feat: i created progress circle and included inside the progress bar so renamed into progress
